Reorganized the code
