+++ src/constants/frontend.ts
@@ -1,0 +1,2 @@
+
+export const backendRefeshTime = 1000;  // 1 second
